update for 'trans -> transform' rename in ggplot2 3.5.0
ggplot2 3.5.0 [*] renamed the 'trans' argument of scale functions to
'transform', marking the old name as deprecated.  That change leads to
a handful of pmplots failures related to the trans default being
lifestyle::deprecated() rather a string.

pmplots interacts with trans in various ways:

 1) internal code uses trans and passes it to ggplot2 scale functions

 2) user-specified trans values can come in via defx() and defy()

 3) user-specified trans values can come in via the xs/ys list
    arguments of dv_pred(), y_time(), and dv_time()

 4) user-specified values can come in via the trans argument of
    pm_log()

Trying to update these cases in a way that's compatible with ggplot2
3.5.0 and earlier versions likely adds a good amount of complication.
Instead follow what pmplots 0.3.6 did and bump the minimum ggplot2
version.

Then deal with *1* by simply using transform instead of trans.  Handle
*2-4* by detecting trans and remapping it to transform, warning the
user that trans is deprecated.

[*] 000a9392e (Rename `trans` arguments to `transform`, 2023-12-14)