test-utils.R: drop as_img helper in favor of list comparison
The R 3.6 and R 4.0 jobs where failing when comparing the rot_xy()
result with a reference result.  cf92820 and 1142477 tried to fix this
by setting the top-level plot_env component of the results to NULL.
That partially fixed the problem but didn't account for plot_env
components that can be deeper in the results.

Later commits switched to comparing a dump of the image.  Revert those
changes, and instead compare the list result after clobbering plot_env
environments at any level of the results.