ci: simplify condition for setting upgrade
Jobs that pin package versions set upgrade=FALSE for the
r-lib/actions/setup-r-dependencies step to prevent pak from pulling in
newer package versions.

The condition for setting the value is already long and is about to
grow.  Make the conditional more readable (and the purpose a little
clearer) by setting upgrade based on a field to the job configuration.