DESCRIPTION: add S7 as suggested package to appease R CMD check
With ggplot2 4.0.0, which switched to using S7 classes, R CMD check
complains:

    * checking dependencies in R code ... WARNING
    '::' or ':::' import not declared from: ‘S7’

pmplots code base doesn't make any direct calls to S7, so I'm not sure
what triggers this warning.

Silence it by adding S7 to Suggests.  The only downside that should
have is the installation of an unnecessary package when a user
requests to install ggplot2 < 4.0.0 and all suggested packages.